refactor(shared): unify tenant interceptor behind TENANT_ID_PROVIDER token
