+++ src/Client/Logistics.Angular/projects/customer-portal/src/app/app.config.ts
@@ -7,14 +7,19 @@ import {
 } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 import { provideRouter, withComponentInputBinding, withRouterConfig } from "@angular/router";
-import { getAccessToken, provideApi, provideSpartanHlm } from "@logistics/shared";
+import {
+  getAccessToken,
+  provideApi,
+  provideSpartanHlm,
+  TENANT_ID_PROVIDER,
+  tenantInterceptor,
+} from "@logistics/shared";
 import { I18nService, TENANT_SETTINGS_PROVIDER } from "@logistics/shared/services";
 import { provideTranslateService } from "@ngx-translate/core";
 import { provideTranslateHttpLoader } from "@ngx-translate/http-loader";
 import { provideAuth } from "angular-auth-oidc-client";
 import { authConfig } from "@/core/auth";
-import { tenantInterceptor } from "@/core/interceptors";
-import { CustomerPortalSettingsProvider } from "@/core/services";
+import { CustomerPortalSettingsProvider, TenantContextService } from "@/core/services";
 import { environment } from "@/env";
 import { routes } from "./app.routes";
 
@@ -41,6 +46,8 @@ export const appConfig: ApplicationConfig = {
     provideAppInitializer(() => {
       inject(I18nService).init({ supportedLanguages: ["en"] });
     }),
+
+    { provide: TENANT_ID_PROVIDER, useExisting: TenantContextService },
 
     // Localization provider
     { provide: TENANT_SETTINGS_PROVIDER, useExisting: CustomerPortalSettingsProvider },
