--- src/Client/Logistics.Angular/projects/customer-portal/src/app/core/interceptors/index.ts
@@ -1,1 +1,0 @@
-export * from "./tenant.interceptor";
