+++ src/Client/Logistics.Angular/projects/shared/src/lib/api/interceptors/index.ts
@@ -1,2 +1,3 @@
 export * from "./error-handler.interceptor";
+export * from "./tenant.interceptor";
 export * from "./token-auth.interceptor";
