--- src/Client/Logistics.Angular/projects/tms-portal/src/app/core/interceptors/index.ts
@@ -1,1 +1,0 @@
-export * from "./tenant.interceptor";
