--- src/Client/Logistics.Angular/projects/tms-portal/src/app/core/interceptors/tenant.interceptor.ts
@@ -1,17 +1,0 @@
-import { HttpRequest, type HttpEvent, type HttpHandlerFn } from "@angular/common/http";
-import { inject } from "@angular/core";
-import { Observable } from "rxjs";
-import { TenantService } from "@/core/services/tenant.service";
-
-export function tenantInterceptor(
-  request: HttpRequest<unknown>,
-  next: HttpHandlerFn,
-): Observable<HttpEvent<unknown>> {
-  const tenantService = inject(TenantService);
-  const headers = tenantService.generateTenantHeaders(request.headers);
-
-  request = request.clone({
-    headers: headers,
-  });
-  return next(request);
-}
